DltDump: Implement Split logic
Take the option to split and apply it to the OutputBase. The option /split is used to define the size the output file must exceed when it is split. The output option /output must include a template that describes the file name, either with %CTR% (which is a counter that increments), or with %CDATE%, %CTIME% or %CDATETIME%.

Issue: DOTNET-602

+++ apps/DltDump/code/Domain/IOutputStreamFactory.cs
@@ -18,7 +18,7 @@
         /// Gets or sets the split.
         /// </summary>
         /// <value>The number of bytes each file should grow to before splitting.</value>
-        int Split { get; set; }
+        long Split { get; set; }
 
         /// <summary>
         /// Creates the output stream based on the output format.
